Move classes to specialized packages
